Add files via upload
Artemis Financial is a financial firm looking to create financial services for their consumers to choose from a different plan according to the consumer needs. They are also looking to incorporate security to ensure that the information being transferred is secure, and there isn't unauthorized access. I enjoyed running dependency testing to view vulnerabilities. 
With this in mind, code must be written securely because it provides less of a chance to have weak areas that allow for hackers to tear passed the security wall. Software security adds credibility to a company's overall well being. When you have good trust with a company, it attracts more clients. 
There was only one thing about the vulnerability assessment that I found a bit challenging, which was identifying the false positives. In order to increase the layers of security I created an SSL certificate as well as making sure the project has HTTP posts. 
Considering the future, I think one strategy I would apply to assess vulnerabilities and determine mitigation techniques finding and figuring out any false positives.  In order to ensure the code and software application were function and secure, I had to run the project a series of times, as well as running as dependency checks. Upon refactoring code, I had to check run the project in order to see any new vulnerabilities. 
The most help resource utilized for the future would be viewing projects in the web browser in eclipse. One particular assignment I would like to showcase is my Vulnerability Assessment Report, specifically my understanding of how to find and understand the information.